+++ util/utf8string.cpp
@@ -267,21 +267,21 @@ void encodeUTF8(uint32_t code, vector<uint8_t>& chars)
         chars.push_back(static_cast<uint8_t>(code));
     } else if (code <= 0x07FF) {
         code -= 0x80;
-        uint8_t c0 = (code >> 6) + 0xc2;
+        uint8_t c0 = static_cast<uint8_t>((code >> 6) + 0xc2);
         uint8_t c1 = (code & 0x3f) + 0x80;
         chars.push_back(c0);
         chars.push_back(c1);
     } else if (code <= 0x0FFF) {
         code -= 0x800;
         uint8_t c0 = 0xe0;
-        uint8_t c1 = (code >> 6) + 0xa0;
+        uint8_t c1 = static_cast<uint8_t>((code >> 6) + 0xa0);
         uint8_t c2 = (code & 0x3f) + 0x80;
         chars.push_back(c0);
         chars.push_back(c1);
         chars.push_back(c2);
     } else if (code <= 0xCFFF) {
         code -= 0x1000;
-        uint8_t c0 = (code >> 12) + 0xe1;
+        uint8_t c0 = static_cast<uint8_t>((code >> 12) + 0xe1);
         uint8_t c1 = ((code >> 6) & 0x3f) + 0x80;
         uint8_t c2 = (code & 0x3f) + 0x80;
         chars.push_back(c0);
@@ -299,7 +299,7 @@ void encodeUTF8(uint32_t code, vector<uint8_t>& chars)
         throw runtime_error("should not reach here. surrogate region");
     } else if (code <= 0xFFFF) {
         code -= 0xE000;
-        uint8_t c0 = (code >> 12) + 0xee;
+        uint8_t c0 = static_cast<uint8_t>((code >> 12) + 0xee);
         uint8_t c1 = ((code >> 6) & 0x3f) + 0x80;
         uint8_t c2 = (code & 0x3f) + 0x80;
         chars.push_back(c0);
@@ -308,7 +308,7 @@ void encodeUTF8(uint32_t code, vector<uint8_t>& chars)
     } else if (code <= 0x3FFFF) {
         code -= 0x10000;
         uint8_t c0 = 0xf0;
-        uint8_t c1 = (code >> 12) + 0x90;
+        uint8_t c1 = static_cast<uint8_t>((code >> 12) + 0x90);
         uint8_t c2 = ((code >> 6) & 0x3f) + 0x80;
         uint8_t c3 = (code & 0x3f) + 0x80;
         chars.push_back(c0);
@@ -317,7 +317,7 @@ void encodeUTF8(uint32_t code, vector<uint8_t>& chars)
         chars.push_back(c3);
     } else if (code <= 0xFFFFF) {
         code -= 0x40000;
-        uint8_t c0 = (code >> 18) + 0xf1;
+        uint8_t c0 = static_cast<uint8_t>((code >> 18) + 0xf1);
         uint8_t c1 = ((code >> 12) & 0x3f) + 0x80;
         uint8_t c2 = ((code >> 6) & 0x3f) + 0x80;
         uint8_t c3 = (code & 0x3f) + 0x80;
@@ -328,7 +328,7 @@ void encodeUTF8(uint32_t code, vector<uint8_t>& chars)
     } else if (code <= 0x10FFFF) {
         code -= 0x100000;
         uint8_t c0 = 0xf4;
-        uint8_t c1 = (code >> 12) + 0x80;
+        uint8_t c1 = static_cast<uint8_t>((code >> 12) + 0x80);
         uint8_t c2 = ((code >> 6) & 0x3f) + 0x80;
         uint8_t c3 = (code & 0x3f) + 0x80;
         chars.push_back(c0);
